+++ tests/testthat/test-mark_nodes.R
@@ -57,3 +57,19 @@ test_that("additional node mark functions work", {
   expect_equal(as.character(node_is_random(ison_adolescents, 1)),
                c("FALSE", "FALSE",  "FALSE",  "TRUE", "FALSE", "FALSE", "FALSE", "FALSE"))
 })
+
+test_that("node infection, exposure, and recovery works", {
+  skip_on_cran()
+  skip_on_ci()
+  set.seed(1234)
+  .data <- play_diffusion(generate_random(to_undirected(to_unnamed(ison_networkers)), with_attr = FALSE),
+                          seeds = 10, latency = 0.25, recovery = 0.2, steps = 10)
+  expect_true(c(node_is_infected(.data, time = 0))[dplyr::filter(summary(.data),
+                                                                 event == "I" & t == 0)$nodes])
+  expect_true(c(node_is_recovered(.data, time = 4))[dplyr::filter(summary(.data),
+                                                                  event == "R" & t == 4)$nodes[1]])
+  expect_true(c(node_is_latent(.data, time = 2))[dplyr::filter(summary(.data),
+                                                               event == "E" & t == 2)$nodes[1]])
+})
+
+